public/python_scripts/requirements.txt
matplotlib==3.7.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
mysql-connector-python-rf==2.2.2
numpy==1.25.2
oauthlib==3.2.2
olefile==0.46openai==0.28.1
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.0.0
paramiko==3.3.1
parse==1.19.1
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfminer3k==1.3.4
pdfplumber==0.9.0
pdfquery==0.4.3
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
pipreqs==0.4.13
platformdirs==3.2.0
playwright==1.32.1
plotly==5.15.0
ply==3.11
posthog==3.0.2
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.4
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==1.10.12
pydeck==0.8.1b0
pydivert==2.1.0
pyee==9.0.4
pyemd==1.0.0
Pygments==2.15.1
pyhtml2pdf==0.0.6
pymdown-extensions==10.3
Pympler==1.0.1
PyMuPDF==1.22.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pypdf==3.15.5
PyPDF2==3.0.1
pypdfium2==4.20.0
PyPika==0.48.9
pyppeteer==1.0.2
pyquery==2.0.0
pyreadline3==3.4.1
pysftp==0.2.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==25.0.2
referencing==0.30.2
regex==2023.6.3
replicate==0.11.0
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.6.0
roman==4.0
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.11.2
selenium-wire==5.1.0
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
slate3k==0.5.3
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.8.1
SQLAlchemy==1.4.49
st-annotated-text==4.0.1
stack-data==0.6.2
starlette==0.27.0
stop-words==2018.7.23
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabula-py==2.7.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.17.1
textract==1.6.5
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.32.1
trio==0.22.0
trio-websocket==0.10.2
TwoCaptcha==0.0.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
undetected-chromedriver==3.5.2
unstructured==0.10.18
watchfiles==0.20.0
wcwidth==0.2.6
webdriver-manager==3.8.6
webdrivermanager==0.10.0
websockets==10.4
Werkzeug==2.2.3
wget==3.2

